Refactor UploadRunner subclasses into StorageBroker subclass, to de-couple generic storage operations from the upload step
