+++ aodncore/testlib/__init__.py
@@ -4,14 +4,14 @@
 from .basetest import BaseTestCase
 from .dummyhandler import DummyHandler
 from .handlertest import HandlerTestCase
-from .testutil import (NullUploadRunner, dest_path_testing, get_nonexistent_path, get_test_config, make_test_file,
+from .testutil import (NullStorageBroker, dest_path_testing, get_nonexistent_path, get_test_config, make_test_file,
                        make_zip, mock)
 
 __all__ = [
     'BaseTestCase',
     'DummyHandler',
     'HandlerTestCase',
-    'NullUploadRunner',
+    'NullStorageBroker',
     'dest_path_testing',
     'get_nonexistent_path',
     'get_test_config',
